Add setting to toggle global hotkey hook
MusicBee already offers a way to set a hotkey to be global and this conflicts with it.

If the hotkey isn't set to global within MusicBee sometimes it doesn't have proper "focus" and the media keys won't work - however, when having them global and using the hook from this plugin, it will send a play and pause close to one another that the player breaks.